Add Click commands and Makefile

+++ ingestion/datasets/dataset.py
@@ -35,10 +35,19 @@ class Dataset():
       members=members
     )
       
-  def process_dataset(self):
-    cached_file_md5 = self.duck_client.get_dataset_cache_md5(self.dataset_config.file_url)
+  def process_dataset(self, check_cached_files_task=True, download_files_task=True, force_download_files=False, unzip_files_task=True, drop_stg_table_task=True, ingest_from_files_task=True):
+    cached_files_changed=False
     online_file_md5 = get_file_checksum_from_https(self.dataset_config.file_url)
-    if cached_file_md5 != online_file_md5 or os.path.exists(self.cache_file_path) == False:
+    
+    if check_cached_files_task:
+      cached_file_md5 = self.duck_client.get_dataset_cache_md5(self.dataset_config.file_url)
+      cached_files_changed = cached_file_md5 != online_file_md5
+    else:
+      logger.info(f'Skip check_cached_files_task')
+      
+    if force_download_files or (
+      download_files_task and (cached_files_changed or os.path.exists(self.cache_file_path) == False)
+    ):
       try: 
         self.download_file_from_https()
         self.duck_client.insert_cache_log(
@@ -51,17 +60,25 @@ class Dataset():
     else:
       logger.info(f'Skipping download for dataset {self.dataset_name}.')
     
-    if Path(self.dataset_config.cache_file_name).suffix == ".zip":
+    if unzip_files_task and Path(self.dataset_config.cache_file_name).suffix == ".zip":
       self.unzip_files()
-
-    logger.info(f'Drop stg table {self.dataset_config.stg_table}')
-    self.duck_client.drop_table('stg', self.dataset_config.stg_table)
+    else:
+      logger.info(f'Skip unzip_files_task')
+      
+    if drop_stg_table_task:
+      logger.info(f'Drop stg table {self.dataset_config.stg_table}')
+      self.duck_client.drop_table('stg', self.dataset_config.stg_table)
+    else:
+      logger.info(f'Skip drop_stg_table_task')
     
-    self.duck_client.ingest_from_files(
-      files_type=self.dataset_config.file_type,
-      table_name=self.dataset_config.stg_table,
-      cache_path=self.cache_folder_path,
-      cache_sub_folder=self.dataset_config.cache_sub_folder
-    )
+    if ingest_from_files_task:
+      self.duck_client.ingest_from_files(
+        files_type=self.dataset_config.file_type,
+        table_name=self.dataset_config.stg_table,
+        cache_path=self.cache_folder_path,
+        cache_sub_folder=self.dataset_config.cache_sub_folder
+      )
+    else:
+      logger.info(f'Skip ingest_from_files_task')
       
     return True
